[Workaround] panfrost: Add debug flag to flush batches on new framebuffer
Some games (such as Monolith) use the bound framebuffer as a texture
after rendering to it, though doing that is undefined behaviour.

To workaround this in some cases, add a debug flag which only allows a
single batch to be active at a time, so that when the texture fetch
reads the old pixel values from memory, the batch has been flushed and
they have been updated.

AFBC can still cause artefacts, so noafbc should be used with this.