[Firefox][Spill][Bug] pan/mdg: Fix mask usage when filling before a spill
Check the bytemask against 0xFFFF rather than 0xF so that the fill is
skipped for a .xyzw write rather than a .x write.

Set the mask on the store to 0xF when doing a read so that all
components are written back.

TODO: Should write_mask be initialised based on the bytemask rather
than the raw mask?