[#5839][Bug] panfrost: Update viewport before deciding to skip rasterization
panfrost_emit_viewport sets the value of scissor_culls_everything,
which is used to decide whether or not to proceed with a draw. But
this function will not be called if skip_rasterization decides not to
continue with the draw, and so after one draw is skipped, every
following one is skipped until the end of the batch.

To fix this, call panfrost_emit_viewport if we need to before checking
whether to skip the draw. To avoid repeating this needlessly if there
are many skipped draws in a row, clear the dirty bits from
ctx->dirty. However, other code relies on the dirty bits still being
set, so add new dirty flags and set both when marking the state dirty.

(The odd ordering of the bits in the dirty enum, with the two DSC
entries next to each other, saves a couple of mov instructions on
AArch64, but isn't very important.)

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/5839
Fixes: 79356b2e5f1 ("panfrost: Skip rasterizer discard draws without side effects")