[Tiler] panfrost: Document probable reason for magic 0xa0000000 value
Opcodes between 0xa0000000 and 0xfffffffe set the high bits of the
pointer to the draw struct on later generations. Probably this is the
same for v4 hardware. This indicates that v4 hardware might leak the
pointer between GPU jobs, which isn't a very big problem for
panfrost.ko where the addresses are so predictable anyway.

If the point of this instruction was only to serve as a NOP, then
0xffffffff would have worked as well. It only needs to be written for
the first tile, because the high bits of the pointer will stay reset
for the rest of the job (unless they actually need to be changed..).

But this explanation seems somewhat suspect--the pointer to the end of
the polygon list entry is not updated, so wouldn't the tiler overwrite
it for the first draw that covers that tile?