*[CRCDump][WIP] panfrost: Add support for CPU verification of CRCs
The polynomial used is from ECMA-182, and is also used for a very
similar CRC implementation in XZ, except that XZ inverts the CRC value
before returning it to make it easier to chain invocations of the CRC
function.

Proof:

A 16x16 RGBA tile with all transparent black except that the first
pixel (in memory order) has a red value of 64 gives a CRC of
0x633ddb30f6b0286b.

Inverted, this is 0x9cc224cf094fd794.

$ dd if=/dev/zero of=data.bin bs=1024 count=1
$ echo @ | dd conv=notrunc of=data.bin bs=1 count=1
$ xz data.bin
$ od -t x1 data.bin.xz | cut -s -d" " -f2- | tr " " "\n " |
> tac | tr -d "\n" | grep -c 9cc224cf094fd794
1

panfrost_calculate_crc is similar to the crc64_init function in the XZ
source, but we do not use a lookup table as speed is not very
important for this feature (yet..).

TODO: Calculate AFBC subtiling swizzles rather than using a table? But
any calculations will likely use more space than the table itself…