[BO Dump] panfrost: Emit fragment job before submitting vertex job
We want to share the list of BOs between vertex and fragment jobs.
Emitting the fragment job may allocate memory, so move this to be done
earlier.

(Eventually we may want to only add BOs to jobs when they are used by
the specified fragment/vertex stage, but even then it will probably be
faster to collect all potential BOs once per batch, at least while
the (arguably) poor decision of using a sparse_array is in place.)