panfrost: Add non command-stream v10 XML
Add the structs for the v10 XML which do not need command-stream
packing.

Notable changes from v9:

- CRC Clear Color is now only 48 bits rather than 64. (The underlying
  polynomial changes in accordance.)

- Some pointers have been changed from a width of 64 bits to 48 bits,
  to be compatible with the corresponding CS structs.

There may be other changes, many of the fields have not been verified.

The commented fields in the Frame Draw struct are likely to be
correct, but are removed until the corresponding fields in the CS Draw
struct are confirmed.

(It's possible that pointers are still 64 bits even for CS structs,
but in practice VAs will be 48-bit and it's more efficient to upload
the reduced size.)