pan/gen_pack: Add macros to conditionally use cs pack functions
For structs that used to be packed normally but now use command-stream
packing with v10, create a single macro that can handle both
situations, by passing in the arguments required for both cases.

The _cs_v10 postfix is rather long, and maybe could be shortened
because it is unlikely that v11 will convert more structs to
command-stream specification.