panfrost: Add v10 command-stream structs
A number of fields are unknown, but enough of them are included to be
useful.

Note that start positions for command-stream registers are all
hex--this is partly because they were originally taken from hexdumps,
and partly to make it obvious what layout a struct is using.

Logically there are 256 32-bit words where values can be written, but
the actual supported number of registers can be queried with the macro
STREAM_FEATURES_WORK_REGISTERS_GET in the kbase source. As an example,
G610 sets this to 95.

Many pointer fields are 48-bit rather than 64-bit; this reduces the
command-stream size for each from sixteen bytes to eight.