An observation:
When two compute shaders run at the same time, writes to the same
cache line can be lost, even if atomic accesses are used (!).

It would appear that L1 caches are not coherent between cores???

  ("cores": That would be odd, since presumably normal compute jobs
  can split workgroups between cores... right? I'm actually not sure.
  Tiles in fragment jobs are at least handled by single cores.)

Is there a way for atomic operations to be atomic wrt more than just
the same core?